rename folder test for servers
